LIBFCREPO-766. Removed hard-coded "/apps/git" directory from Vagrantfile
Modified the Vagrantfile to not assume that the repository has been
checked out into the "/apps/git" directory. The Vagrantfile now
uses relative paths to access other repositories needed for generating
the Vagrant machines.

Updated the README.md

Added "bootstrap-data" and "umd-fcrepo-docker" to .gitignore as these
are commonly copied into into the "fcrepo-vagrant" directory, but
should not be checked in to the "fcrepo-vagrant" repository.

https://issues.umd.edu/browse/LIBFCREPO-766